HEEDLS-532 Review markups
