+++ DigitalLearningSolutions.Web/Helpers/FilterableTagHelper.cs
@@ -2,6 +2,7 @@
 {
     using System.Collections.Generic;
     using DigitalLearningSolutions.Data.Models.User;
+    using DigitalLearningSolutions.Web.Helpers.FilterOptions;
     using DigitalLearningSolutions.Web.ViewModels.Common.SearchablePage;
 
     public static class FilterableTagHelper
